refactor(siteImport): one adapter contract, phase-decomposed pipeline, deduped helpers
The site importer was the worst-quality module in the repo by every audit
metric (applyImport.ts: 700 LOC, cognitive complexity 178, worst MI of any
production file) while being one of the product's most important features.

- ONE transaction contract. The editor store re-declared the entire
  SiteImportTransaction interface as `SuperImportHelpers` (87 identical
  lines, hand-maintained twice). The store now implements
  `SiteImportTransaction` from @core/siteImport directly; the duplicate
  interface is deleted.
- applyImport.ts split along its two halves:
  - buildPlan.ts — buildImportPlan as straight-line phase calls; each phase
    (HTML page plans, stylesheet partition, @import expansion, CSS parse,
    cross-sheet classes, asset plan, conflicts) is a named function.
  - commitPlan.ts — commitImportPlan with one named function per commit
    concern (uploads, Google fonts, page-id minting, tokens, fonts, rules,
    pages, page-scoped files).
  - planCss.ts — the ONE CSS-source parse path. The external-stylesheet and
    inline-<style> loops were 28 copy-pasted lines each; both now call
    parseCssSourceIntoPlan, so a new token extractor or warning kind can
    never silently miss one path.
- paths.ts — dirname/joinPaths existed verbatim in both assetPlan.ts and
  htmlPagePlan.ts (the comment even admitted the duplication). One copy now.
- readCssDeclarationBag in @core/css-substitution — the CSSOM→camelCase-bag
  walker (substitution decode + custom-property preservation + security
  gate) existed in both cssToStyleRules.parseDeclarations and
  htmlImport's extractInlineStyles. One walker now; both wrap it.
- module-size-budgets ratchet lowered for the shrunken cssToStyleRules.ts;
  docs/features/site-import.md layout updated.

Complexity (fallow): applyImport (cog 178, MI 75.8) → buildPlan (57/85.2)
+ commitPlan (66/86.8) + planCss (13/88.4). Behavior preserved — 5,343
tests pass, including the 389 import-suite tests.
